pyproject.toml
[project]
name = "cryptopulse"
version = "0.1.0"
description = "Web-API for tracking cryptocurrencies"
authors = [
    {name = "Nicolas Miroshnikov",email = "shoveit808@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "bcrypt (>=4.0.1,<4.1.1)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic[email] (>=2.11.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
